tts_requirement.txt
anyio
boto3
botocore
cachetools
certifi
charset-normalizer
click
fastapi
google-api-core
google-auth
google-cloud-texttospeech
googleapis-common-protos
grpcio
grpcio-status
gunicorn
h11
idna
importlib-metadata
jmespath
proto-plus
protobuf
pyasn1
pyasn1-modules
pydantic
python-dateutil
requests
rsa
s3transfer
setuptools
six
sniffio
starlette
typing_extensions
urllib3
uvicorn
zipp    